Creating final amount simulation and improving action returns

+++ FinancialGoalManager.Application/Commands/FinancialGoals/DeleteGoal/DeleteGoalCommandHandler.cs
@@ -4,7 +4,7 @@ using MediatR;
 
 namespace FinancialGoalManager.Application.Commands.FinancialGoals.DeleteGoal
 {
-    public class DeleteGoalCommandHandler : IRequestHandler<DeleteGoalCommand>
+    public class DeleteGoalCommandHandler : IRequestHandler<DeleteGoalCommand, bool>
     {
         private readonly IFinancialGoalRepository _goalRepository;
         public DeleteGoalCommandHandler(IFinancialGoalRepository goalRepository)
@@ -12,11 +12,16 @@ namespace FinancialGoalManager.Application.Commands.FinancialGoals.DeleteGoal
             _goalRepository = goalRepository;
         }
 
-        public async Task Handle(DeleteGoalCommand request, CancellationToken cancellationToken)
+        public async Task<bool> Handle(DeleteGoalCommand request, CancellationToken cancellationToken)
         {
             var financialGoalDto = await _goalRepository.GetGoalsById(request.Id);
 
+            if (financialGoalDto == null)
+                return false;
+
             await _goalRepository.DeleteGoal(financialGoalDto);
+
+            return true;
         }
     }
 }
